Rename agency reducer to agencies
